+++ src/common/pico_util/queue.c
@@ -11,8 +11,8 @@
 void queue_init_with_spinlock(queue_t *q, uint element_size, uint element_count, uint spinlock_num) {
     q->lock = spin_lock_instance(spinlock_num);
     q->data = (uint8_t *)calloc(element_count + 1, element_size);
-    q->element_count = element_count;
-    q->element_size = element_size;
+    q->element_count = (uint16_t)element_count;
+    q->element_size = (uint16_t)element_size;
     q->wptr = 0;
     q->rptr = 0;
 }
